+++ src/extension-test/unit/cursor-doc/paredit-test.ts
@@ -2166,9 +2166,33 @@ describe('paredit', () => {
       });
 
       describe('Slurping backwards', () => {
-        it('slurps form before string', async () => {
+        it('slurps form before non-empty string', async () => {
           const a = docFromTextNotation('(str) "fo|o"');
           const b = docFromTextNotation('"(str) fo|o"');
+          await paredit.backwardSlurpSexp(a);
+          expect(textAndSelection(a)).toEqual(textAndSelection(b));
+        });
+        it('slurps form before empty string without adding trailing space', async () => {
+          const a = docFromTextNotation('foo "|"');
+          const b = docFromTextNotation('"foo|"');
+          await paredit.backwardSlurpSexp(a);
+          expect(textAndSelection(a)).toEqual(textAndSelection(b));
+        });
+        it('slurps form before empty list without adding trailing space', async () => {
+          const a = docFromTextNotation('foo (|)');
+          const b = docFromTextNotation('(foo|)');
+          await paredit.backwardSlurpSexp(a);
+          expect(textAndSelection(a)).toEqual(textAndSelection(b));
+        });
+        it('slurps form before whitespace-only list without adding trailing space', async () => {
+          const a = docFromTextNotation('foo (|   )');
+          const b = docFromTextNotation('(foo|)');
+          await paredit.backwardSlurpSexp(a);
+          expect(textAndSelection(a)).toEqual(textAndSelection(b));
+        });
+        it('slurps form before empty vector without adding trailing space', async () => {
+          const a = docFromTextNotation('foo [|]');
+          const b = docFromTextNotation('[foo|]');
           await paredit.backwardSlurpSexp(a);
           expect(textAndSelection(a)).toEqual(textAndSelection(b));
         });
@@ -2180,9 +2204,7 @@ describe('paredit', () => {
         });
         it('slurps form before list including meta and readers', async () => {
           const a = docFromTextNotation('^{:a b} #c ^d "foo" (|)');
-          // TODO: Figure out how to test result after format
-          //       (Because that last space is then removed)
-          const b = docFromTextNotation('(^{:a b} #c ^d "foo" |)');
+          const b = docFromTextNotation('(^{:a b} #c ^d "foo"|)');
           await paredit.backwardSlurpSexp(a);
           expect(textAndSelection(a)).toEqual(textAndSelection(b));
         });