+++ CHANGELOG.md
@@ -4,8 +4,8 @@ Changes to Calva.
 
 ## [Unreleased]
 
-- Fix: Slurp Forward no longer adds leading space when slurping into empty forms. [#1440](https://github.com/BetterThanTomorrow/calva/issues/1440) &
-[#2085](https://github.com/BetterThanTomorrow/calva/issues/2085) 
+- Fix: Slurp Forward no longer adds leading space when slurping into empty forms. [#1440](https://github.com/BetterThanTomorrow/calva/issues/1440) & [#2085](https://github.com/BetterThanTomorrow/calva/issues/2085)
+- Fix: Slurp Backward no longer adds trailing space when slurping into empty forms
 
 ## [2.0.546] - 2026-01-30